feat: Move workload templates to files and add serve option
This change adds a `serve` flag to the push command. The option can be
used to specify which inference service to use for deploying the model.

This change also moves the workload template from code to files under
the embedded-files directory. This makes adding new templates and
making changes to existing ones easier.