feat: Add new tasks to the tekton pipeline
This change adds two new tasks to the tekton pipeline:

- get-serving-workload: this task finds the workload type that will
be used for deploying the model
- model-pkl-to-joblib: this task will create a copy of the model file
with the name `model.joblib` which is required for some inferece services
such as kfserving.

This PR also changes `train` task to, instead of saving the RUN_ID which
was used by the serve task to get the model uri, get the model-uri and
save it. In that way it can be used directly by the `model-pkl-to-joblib`
and `serve` task.